Replace Vorpal with Heavy (bitdizzy, hellmonk)
Vorpal is a pretty bland brand. It's just more damage, everywhere,
all the time. It's often fine but never exciting.

So let's swap it out for something else I wanted to add anyway.
Heavy is a new brand that increases weapon delay by 50% (the opposite
of Speed), but increases weapon base damage by 75%. (This then gets
multiplied by stats, skills, etc.) It works differently from Vorpal,
Flaming etc because, if it was a post-AC effect, it would be utterly
terrifying in monsters' hands. It's also much easier to display to
players as a pre-AC effect - it'll show up properly in damage rates.

Various unrands have been adjusted as appropriate. Currently the only
one that's keeping Heavy is Sniper (with some tweaking to make it vaguely
usable), but it feels like there should be another good Heavy unrand
design out there somewhere. (Dark Maul gets Freezing instead, since I
really want it to still have a ridiculous delay even at max skill, and
Heavy doesn't do that.)

The idea is that this should feel fun to use (big impacts per keypress),
but should also be a bit scary (potentially multiple enemy actions per
attack). It's very good against monsters with high AC, but quite bad if
you're depending on auxes or slaying for damage. It's a land of contrasts.

Let's try it out!